2 decimal places for pension and percentage

+++ tax-tracker/src/App.js
@@ -17,7 +17,7 @@ const App = () => {
     studentLoan: STUDENT_LOAN_TYPES.includes(queryStrings.studentLoan)
       ? queryStrings.studentLoan
       : undefined,
-    pension: parseInt(queryStrings.pension) || 0,
+    pension: parseFloat(queryStrings.pension) || 0,
     multiplier: queryStrings.multiplier,
   };
   return (
